Error Label fixed

+++ PatternPixSolution/PatternPixTrainerPanel/Model/Training.cs
@@ -61,7 +61,7 @@ namespace PatternPixTrainerPanel.Model
 
 
         public string FormattedDateTime => $"{Date:dd.MM.yy} {TimeOfDay:hh\\:mm}";
-        public int ErrorsScaled => Errors * 10;
+        public double ErrorsScaled => Errors * 10;
